requirements.txt
streamlit>=1.32.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
numpy
Pillow
tensorflow>=2.12.0